Termine de corregir errores y implementar lo del servidor Express. Es probable que este sea mi ultimo commit. Adios

+++ project/src/pages/editproduct.jsx
@@ -56,8 +56,11 @@ const handleSubmit = async (event) => {
     event.preventDefault();
     let newProductImage;
     if ( image ) {
-        const delRef = ref( getStorage(), 'productImage/' + productImageAddress );
-        deleteObject( delRef );
+
+        if ( productImageAddress === state ) {
+            const delRef = ref( getStorage(), 'productImage/' + productImageAddress );
+            deleteObject( delRef );
+        }
 
         const imgRef = ref( getStorage(), 'productImage/' + state  );
         uploadBytes( imgRef, image );
